fix: code quality fixes from analysis report
- Remove trailing space from commons-compress group ID in build.gradle.kts
  so the CVE resolution override now actually matches
- Remove trailing space from Prometheus metric counter name in Metrics.kt
- Restrict Serviceutils logger to @PublishedApi internal (was public val)
- Fix typo ENDRE_HOVESTOL -> ENDRE_HOVEDSTOL in SkeClient.kt
- Replace var haltRun with AtomicBoolean in SkeService.kt to fix visibility
  race condition across coroutine thread hops
- Extract magic numbers 5000/1000 as RESEND_DELAY_MS/LARGE_FILE_THRESHOLD
- Fix CancellationException swallowing in StatusService.getMottaksStatus
- Use toIntOrNull() in AvstemmingRouting to prevent NumberFormatException 500
- Fix BigDecimal written via setDouble -> setBigDecimal in RepositoryExtensions
- Wrap PreparedStatement in use{} in KravRepository.insertAllNewKrav to
  prevent DB cursor leak
- Remove signed JWT assertion from log messages and exception in
  MaskinportenAccessTokenProvider

Agent-Logs-Url: https://github.com/navikt/sokos-ske-krav/sessions/1c57c3f7-0d54-4ef7-8bec-c3c60d5976d4

+++ src/main/kotlin/no/nav/sokos/ske/krav/repository/RepositoryExtensions.kt
@@ -89,7 +89,7 @@ object RepositoryExtensions {
                     is Long -> setLong(idx, param)
                     is String -> setString(idx, param)
                     is LocalDate -> setDate(idx, Date.valueOf(param))
-                    is BigDecimal -> setDouble(idx, param.toDouble())
+                    is BigDecimal -> setBigDecimal(idx, param)
                 }
             }
         }
